Refactor and remove old operators
