--- aioreactive/operators/__init__.py
@@ -1,30 +1,0 @@
-# from .catch_exception import catch_exception
-# from .concat import concat
-# from .debounce import debounce
-# from .delay import delay
-# from .distinct_until_changed import distinct_until_changed
-# from .empty import empty
-# from .filter import filter
-# from .filteri import filteri
-# from .flat_map import flat_map
-# from .from_async_iterable import from_async_iterable
-from .from_iterable import from_iterable
-from .map import map, map_async
-from .transform import transform
-
-# from .max import max
-# from .merge import merge
-# from .never import never
-# from .retry import retry
-# from .scan import scan
-# from .skip import skip
-# from .skip_last import skip_last
-# from .slice import slice
-# from .switch_latest import switch_latest
-# from .take import take
-# from .take_last import take_last
-# from .to_async_iterable import to_async_iterable
-# from .unit import unit
-# from .with_latest_from import with_latest_from
-
-__all__ = ["from_iterable", "map", "map_async", "transform"]
